feat: [website-test-breakdown-002] - [Extract app-shell import, export, locale, and toolbar coverage from App.test.tsx]
