+++ docs/internal/processes/development-guide-relevant-files.md
@@ -168,6 +168,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `ui/src/features/timeline/state/timeline/replayCompletion.ts` | Timeline replay completion projection seam | Dispatch-response replay should keep `WorldCompletion -> DashboardTraceDispatch` shaping and any supported legacy completion fallback reads centralized here instead of reintroducing alias-only wrapper helpers. |
 | `ui/src/features/timeline/state/factoryTimelineStore.test.ts` | Focused timeline replay contract surface | Replay cleanup work should prove canonical and supported legacy dispatch request or response behavior here by asserting reconstructed workstation requests, trace dispatches, runtime sessions, and world state rather than source-layout details. |
 | `ui/src/App.current-selection.test.tsx` | Dedicated current-selection app regression surface | App-level selected-work, trace-card, responsive-layout, terminal-state, and mixed inference/script dispatch-history regressions should live here once the focused current-selection flow would otherwise push `ui/src/App.test.tsx` past the repository’s file or function size limits; when inference attempts or nested request/response bodies are collapsed by default, this seam must prove the app-level flow by expanding the relevant attempt and body disclosures before asserting payload or diagnostic text. |
+| `ui/src/App.import.test.tsx`, `ui/src/App.export-dialog.test.tsx`, `ui/src/App.export-submit.test.tsx`, `ui/src/App.toolbar-locale.test.tsx`, and `ui/src/testing/app-shell-*.ts*` | Focused app-shell regression seams | When `ui/src/App.test.tsx` grows past the repo size gates, split shell behaviors into focused app-level files by surface area, keep shared render, lifecycle, import, and export probes local to `ui/src/testing/app-shell-*.ts*`, and extend the explicit Biome allowlist with those concrete files instead of reintroducing blanket test exceptions or rebuilding another monolithic app harness. |
 | `ui/src/features/current-selection/workstation-request-detail.tsx` | Workstation-request detail fallback seam | Request-detail rendering, script-versus-inference alternate panels, sparse projected field fallbacks, and the omission of low-value request-count summary rows should stay centralized here so current-selection request diagnostics keep one observable detail contract instead of splitting fallback copy across summary cards or app shells. |
 | `ui/src/features/current-selection/workstation-request-detail.test.tsx` | Workstation-request detail regression surface | Customer-visible request-detail regressions should assert rendered markdown/plain-text prompt handling, hidden request-count summary labels, plus sparse request, trace, script-request, and script-response fallback copy here rather than relying on helper-only tests or broader app smoke coverage. |
 | `ui/src/features/current-selection/inference-attempt.tsx` and `ui/src/features/current-selection/execution-details.test.tsx` | Inference-attempt disclosure seam | Current-selection inference attempts should own their own collapsed-by-default disclosure state here, keep attempt status and concise timing visible in the collapsed summary, and prove per-attempt `aria-expanded` behavior plus independent expansion through the focused execution-details regression surface instead of broader widget smoke tests. |
