+++ ui/src/App.test.tsx
@@ -19,7 +19,6 @@ import type {
 } from "./api/dashboard";
 import type { FactoryEvent } from "./api/events";
 import { FACTORY_EVENT_TYPES } from "./api/events";
-import type { FactoryValue } from "./api/named-factory";
 import {
   buildDashboardSnapshotFixture,
   dashboardWorkstationRequestFixtures,
@@ -42,11 +41,6 @@ import {
   semanticWorkflowDashboardSnapshot,
   twentyNodeDashboardSnapshot,
 } from "./components/dashboard/test-fixtures";
-import {
-  DASHBOARD_BODY_TEXT_CLASS,
-  DASHBOARD_PAGE_HEADING_CLASS,
-  DASHBOARD_SUPPORTING_LABELS_CLASS,
-} from "./components/ui/dashboard-typography";
 import { formatDurationMillis } from "./components/ui/formatters";
 import { reloadDashboardLayoutFromStorage } from "./features/bento";
 import { useDashboardBentoStore } from "./features/bento/state";
@@ -60,10 +54,7 @@ import {
   createDefaultDashboardStreamState,
   useDashboardStreamStore,
 } from "./features/dashboard/state";
-import * as factoryPngExportModule from "./features/export/factory-png-export";
 import { useExportDialogStore } from "./features/export/state";
-import type { FactoryPngImportValue } from "./features/import";
-import * as factoryPngImportModule from "./features/import/factory-png-import";
 import type { WorldState } from "./features/timeline/state";
 import { useFactoryTimelineStore } from "./features/timeline/state";
 import {
@@ -147,8 +138,6 @@ const activeWorkLabel = "Active Story";
 const eventTimelineWorkID = "work-event-story";
 const eventTimelineTraceID = "trace-event-story";
 const eventTimelineDispatchID = "dispatch-event-story";
-const ONE_PIXEL_PNG_BASE64 =
-  "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVQIHWP4////fwAJ+wP9KobjigAAAABJRU5ErkJggg==";
 const fanInResultWorkID = "work-result";
 const fanInResultLabel = "Implemented Story";
 
@@ -222,14 +211,6 @@ const historicalTimelineSnapshot = {
   ...baselineSnapshot,
   tick_count: 1,
 } satisfies DashboardSnapshot;
-const importedFactorySnapshot = (() => {
-  const snapshot = structuredClone(semanticWorkflowDashboardSnapshot);
-
-  snapshot.factory_state = "Imported factory active";
-  snapshot.tick_count = semanticWorkflowDashboardSnapshot.tick_count + 1;
-
-  return snapshot;
-})();
 
 const { edges: _omittedEdges, ...oneNodeTopologyWithoutEdges } =
   oneNodeDashboardTopology;
@@ -1122,78 +1103,6 @@ const tickZeroInitialStructureRequestEvents: FactoryEvent[] = [
   ),
 ];
 
-const exportTimelineEvents: FactoryEvent[] = [
-  factoryEvent("export-run-request", 0, FACTORY_EVENT_TYPES.runRequest, {
-    factory: {
-      id: "semantic-workflow",
-      name: "semantic-workflow",
-      factoryDirectory: "/work/factories/semantic-workflow",
-      workers: [
-        {
-          modelProvider: "CODEX",
-          name: "reviewer",
-          executorProvider: "SCRIPT_WRAP",
-          type: "MODEL_WORKER",
-        },
-      ],
-      workTypes: [
-        {
-          name: "story",
-          states: [
-            { name: "new", type: "INITIAL" },
-            { name: "done", type: "TERMINAL" },
-          ],
-        },
-      ],
-      workstations: [
-        {
-          id: "review",
-          inputs: [{ state: "new", workType: "story" }],
-          name: "Review",
-          onFailure: [{ state: "done", workType: "story" }],
-          outputs: [{ state: "done", workType: "story" }],
-          worker: "reviewer",
-        },
-      ],
-    },
-    recordedAt: "2026-04-16T12:00:00Z",
-  }),
-];
-const currentNamedFactoryExportResponse = {
-  metadata: {
-    contractSource: "current-factory-api",
-  },
-  id: "authored-current-factory",
-  name: "semantic-workflow",
-  workers: [
-    {
-      executorProvider: "SCRIPT_WRAP",
-      modelProvider: "CODEX",
-      name: "reviewer",
-      type: "MODEL_WORKER",
-    },
-  ],
-  workTypes: [
-    {
-      name: "story",
-      states: [
-        { name: "new", type: "INITIAL" },
-        { name: "done", type: "TERMINAL" },
-      ],
-    },
-  ],
-  workstations: [
-    {
-      id: "review",
-      inputs: [{ state: "new", workType: "story" }],
-      name: "Review",
-      onFailure: [{ state: "done", workType: "story" }],
-      outputs: [{ state: "done", workType: "story" }],
-      type: "MODEL_WORKSTATION",
-      worker: "reviewer",
-    },
-  ],
-} satisfies FactoryValue;
 const queryClients: QueryClient[] = [];
 let restoreBrowserTestShims: (() => void) | null = null;
 
@@ -1369,10 +1278,6 @@ function submitWorkCardControls() {
   };
 }
 
-function fromBase64(value: string): Uint8Array {
-  return Uint8Array.from(atob(value), (character) => character.charCodeAt(0));
-}
-
 function jsonResponse(
   body: unknown,
   status = 200,
@@ -1387,99 +1292,6 @@ function jsonResponse(
   });
 }
 
-function toArrayBuffer(bytes: Uint8Array): ArrayBuffer {
-  const copy = new Uint8Array(bytes.byteLength);
-  copy.set(bytes);
-  return copy.buffer;
-}
-
-function exportImageFile(): File {
-  return new File(
-    [toArrayBuffer(fromBase64(ONE_PIXEL_PNG_BASE64))],
-    "cover.png",
-    { type: "image/png" },
-  );
-}
-
-function createDeferredPromise<T>(): {
-  promise: Promise<T>;
-  reject: (reason?: unknown) => void;
-  resolve: (value: T) => void;
-} {
-  let reject!: (reason?: unknown) => void;
-  let resolve!: (value: T) => void;
-  const promise = new Promise<T>((resolvePromise, rejectPromise) => {
-    reject = rejectPromise;
-    resolve = resolvePromise;
-  });
-
-  return {
-    promise,
-    reject,
-    resolve,
-  };
-}
-
-function installExportDownloadProbe(): {
-  getDownloadedBlob: () => Blob | null;
-  getDownloadedFilename: () => string;
-  restore: () => void;
-} {
-  class MockExportOffscreenCanvas {
-    public constructor(
-      public readonly width: number,
-      public readonly height: number,
-    ) {}
-
-    public getContext(_contextID: "2d"): OffscreenCanvasRenderingContext2D {
-      return {
-        drawImage() {},
-      } as unknown as OffscreenCanvasRenderingContext2D;
-    }
-
-    public async convertToBlob(): Promise<Blob> {
-      return new Blob([toArrayBuffer(fromBase64(ONE_PIXEL_PNG_BASE64))], {
-        type: "image/png",
-      });
-    }
-  }
-
-  const originalCreateObjectURL = window.URL.createObjectURL;
-  const originalRevokeObjectURL = window.URL.revokeObjectURL;
-  const originalCreateImageBitmap = globalThis.createImageBitmap;
-  const originalOffscreenCanvas = globalThis.OffscreenCanvas;
-  const originalClick = HTMLAnchorElement.prototype.click;
-  let downloadedBlob: Blob | null = null;
-  let downloadedFilename = "";
-
-  window.URL.createObjectURL = ((blob: Blob) => {
-    downloadedBlob = blob;
-    return "blob:app-test-export";
-  }) as typeof URL.createObjectURL;
-  window.URL.revokeObjectURL = (() => {}) as typeof URL.revokeObjectURL;
-  globalThis.createImageBitmap = (async () => ({
-    close: () => {},
-    height: 1,
-    width: 1,
-  })) as typeof createImageBitmap;
-  globalThis.OffscreenCanvas =
-    MockExportOffscreenCanvas as unknown as typeof OffscreenCanvas;
-  HTMLAnchorElement.prototype.click = function click(): void {
-    downloadedFilename = this.download;
-  };
-
-  return {
-    getDownloadedBlob: () => downloadedBlob,
-    getDownloadedFilename: () => downloadedFilename,
-    restore: () => {
-      window.URL.createObjectURL = originalCreateObjectURL;
-      window.URL.revokeObjectURL = originalRevokeObjectURL;
-      globalThis.createImageBitmap = originalCreateImageBitmap;
-      globalThis.OffscreenCanvas = originalOffscreenCanvas;
-      HTMLAnchorElement.prototype.click = originalClick;
-    },
-  };
-}
 function renderTraceDrilldownHarness({
   selectedWorkID,
   timelineEvents,
@@ -1503,36 +1315,6 @@ function renderTraceDrilldownHarness({
       <TraceDrilldownTestHarness selectedWorkID={selectedWorkID} />
     </QueryClientProvider>,
   );
-}
-
-function createFactoryImportValue(): FactoryPngImportValue {
-  return {
-    factory: {
-      name: "Dropped Factory",
-      workTypes: [],
-      workers: [],
-      workstations: [],
-    },
-    previewImageSrc: "blob:factory-preview",
-    revokePreviewImageSrc: vi.fn(),
-    schemaVersion: "portos.agent-factory.png.v1",
-  };
-}
-
-function createFileDropTransfer(files: File[]): {
-  dataTransfer: {
-    dropEffect: string;
-    files: File[];
-    types: string[];
-  };
-} {
-  return {
-    dataTransfer: {
-      dropEffect: "none",
-      files,
-      types: ["Files"],
-    },
-  };
 }
 
 function _resizeDashboardViewport(width: number): void {
@@ -1726,973 +1508,6 @@ function registerAppFollowUpTestLifecycle(): void {
     vi.unstubAllGlobals();
   });
 }
-
-describe("App shell import and export flows", () => {
-  registerAppDashboardTestLifecycle();
-
-  it("renders the operator graph for an empty runtime snapshot", async () => {
-    renderApp({ snapshot: baselineSnapshot });
-
-    expect(
-      await screen.findByRole("heading", { name: "you-agent-factory" }),
-    ).toBeTruthy();
-    expect(screen.getByRole("heading", { name: "Factory graph" })).toBeTruthy();
-    expect(screen.getByText("In progress")).toBeTruthy();
-    expect(
-      await screen.findByRole("button", { name: "Select Plan workstation" }),
-    ).toBeTruthy();
-    expect(
-      await screen.findByRole("button", {
-        name: "Select Implement workstation",
-      }),
-    ).toBeTruthy();
-    expect(
-      await screen.findByRole("button", { name: "Select Review workstation" }),
-    ).toBeTruthy();
-    expect(screen.queryByText("Idle")).toBeNull();
-    expect(screen.queryByText("Live Workstation Dashboard")).toBeNull();
-    expect(
-      screen.queryByText(
-        /Reconstruction-first workflow graph with live workstation overlays/i,
-      ),
-    ).toBeNull();
-    expect(
-      screen.queryByRole("heading", { name: "Terminal summary" }),
-    ).toBeNull();
-  });
-
-  it("posts the dropped factory import as a direct canonical /factory activation payload", async () => {
-    const file = new File(["png"], "factory-import.png", { type: "image/png" });
-    const importValue = createFactoryImportValue();
-    vi.spyOn(factoryPngImportModule, "readFactoryImportPng").mockResolvedValue({
-      ok: true,
-      value: importValue,
-    });
-    const { fetchMock } = renderApp({
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    fetchMock.mockImplementation(
-      async (input: RequestInfo | URL, init?: RequestInit) => {
-        const path =
-          typeof input === "string"
-            ? input
-            : input instanceof URL
-              ? `${input.pathname}${input.search}`
-              : input.url;
-
-        if (path === "/factory") {
-          return new Response(JSON.stringify(importValue.factory), {
-            headers: {
-              "Content-Type": "application/json",
-            },
-            status: 200,
-          });
-        }
-
-        throw new Error(
-          `unexpected fetch for ${path} (${init?.method ?? "GET"})`,
-        );
-      },
-    );
-
-    const viewport = await screen.findByRole("region", {
-      name: "Work graph viewport",
-    });
-
-    fireEvent.drop(viewport, createFileDropTransfer([file]));
-
-    const previewDialog = await screen.findByRole("dialog", {
-      name: "Review factory import",
-    });
-    expect(previewDialog.textContent).toContain("Dropped Factory");
-    expect(previewDialog.textContent).toContain("factory-import.png");
-
-    fireEvent.click(
-      within(previewDialog).getByRole("button", { name: "Activate factory" }),
-    );
-
-    await waitFor(() => {
-      expect(fetchMock).toHaveBeenCalledWith(
-        "/factory",
-        expect.objectContaining({
-          body: JSON.stringify(importValue.factory),
-          headers: {
-            "Content-Type": "application/json",
-          },
-          method: "POST",
-        }),
-      );
-    });
-  });
-
-  it("smoke tests authored export and dropped import as one dashboard-shell roundtrip", async () => {
-    const exportProbe = installExportDownloadProbe();
-    const mockedExportResult =
-      await factoryPngExportModule.writeFactoryExportPng({
-        factory: currentNamedFactoryExportResponse,
-        image: exportImageFile(),
-        rasterizeImageToPngBytes: async () => fromBase64(ONE_PIXEL_PNG_BASE64),
-      });
-    if (!mockedExportResult.ok) {
-      throw new Error(
-        "expected the roundtrip export fixture to build successfully",
-      );
-    }
-    const writeFactoryExportPngSpy = vi
-      .spyOn(factoryPngExportModule, "writeFactoryExportPng")
-      .mockResolvedValue(mockedExportResult);
-    const { fetchMock } = renderApp({
-      snapshot: semanticWorkflowDashboardSnapshot,
-      timelineEvents: exportTimelineEvents,
-    });
-
-    fetchMock.mockImplementation(
-      async (input: RequestInfo | URL, init?: RequestInit) => {
-        const path =
-          typeof input === "string"
-            ? input
-            : input instanceof URL
-              ? `${input.pathname}${input.search}`
-              : input.url;
-
-        if (
-          path ===
-          `/factories/${DEFAULT_FACTORY_SESSION_ID}/factory/~current`
-        ) {
-          return jsonResponse(currentNamedFactoryExportResponse);
-        }
-
-        if (path === "/factory") {
-          return jsonResponse(JSON.parse(String(init?.body)));
-        }
-
-        throw new Error(
-          `unexpected fetch for ${path} (${init?.method ?? "GET"})`,
-        );
-      },
-    );
-
-    try {
-      fireEvent.click(
-        await screen.findByRole("button", { name: "Export PNG" }),
-      );
-
-      const exportDialog = await screen.findByRole("dialog", {
-        name: "Export factory",
-      });
-      await waitFor(() => {
-        expect(
-          (
-            within(exportDialog).getByRole("button", {
-              name: "Export PNG",
-            }) as HTMLButtonElement
-          ).disabled,
-        ).toBe(false);
-      });
-
-      const imageInput = within(exportDialog).getByLabelText(
-        "Cover image",
-      ) as HTMLInputElement;
-      Object.defineProperty(imageInput, "files", {
-        configurable: true,
-        value: [exportImageFile()],
-      });
-      fireEvent.change(imageInput);
-      fireEvent.click(
-        within(exportDialog).getByRole("button", { name: "Export PNG" }),
-      );
-
-      await waitFor(() => {
-        expect(exportProbe.getDownloadedBlob()).not.toBeNull();
-      });
-      await waitFor(() => {
-        expect(exportProbe.getDownloadedFilename()).toBe(
-          "semantic-workflow.png",
-        );
-      });
-      await waitFor(() => {
-        expect(within(exportDialog).getByRole("status").textContent).toContain(
-          "Downloaded semantic-workflow.png.",
-        );
-      });
-      fireEvent.click(
-        within(exportDialog).getByRole("button", { name: "Close" }),
-      );
-      await waitFor(() => {
-        expect(
-          screen.queryByRole("dialog", { name: "Export factory" }),
-        ).toBeNull();
-      });
-
-      const exportedBlob = exportProbe.getDownloadedBlob();
-      if (!(exportedBlob instanceof Blob)) {
-        throw new Error("expected the export flow to download a PNG blob");
-      }
-
-      const viewport = await screen.findByRole("region", {
-        name: "Work graph viewport",
-      });
-      fireEvent.drop(
-        viewport,
-        createFileDropTransfer([
-          new File([exportedBlob], exportProbe.getDownloadedFilename(), {
-            type: "image/png",
-          }),
-        ]),
-      );
-
-      const previewDialog = await screen.findByRole("dialog", {
-        name: "Review factory import",
-      });
-      expect(previewDialog.textContent).toContain(
-        currentNamedFactoryExportResponse.name,
-      );
-      expect(previewDialog.textContent).toContain("semantic-workflow.png");
-
-      fireEvent.click(
-        within(previewDialog).getByRole("button", { name: "Activate factory" }),
-      );
-
-      await waitFor(() => {
-        const activationCall = fetchMock.mock.calls.find(
-          ([url]) => url === "/factory",
-        );
-        expect(activationCall).toBeDefined();
-        expect(activationCall?.[1]).toEqual(
-          expect.objectContaining({
-            body: expect.any(String),
-            headers: {
-              "Content-Type": "application/json",
-            },
-            method: "POST",
-          }),
-        );
-        expect(JSON.parse(String(activationCall?.[1]?.body))).toEqual(
-          currentNamedFactoryExportResponse,
-        );
-      });
-      await waitFor(() => {
-        expect(MockEventSource.instances.length).toBeGreaterThan(0);
-      });
-
-      const refreshedStream = MockEventSource.instances.at(-1);
-      if (!refreshedStream) {
-        throw new Error("expected a dashboard stream after factory activation");
-      }
-
-      act(() => {
-        refreshedStream.emit("snapshot", importedFactorySnapshot);
-      });
-
-      await waitFor(() => {
-        expect(screen.queryByText("Imported factory active")).toBeNull();
-      });
-      expect(
-        await screen.findByRole("button", {
-          name: "Select Review workstation",
-        }),
-      ).toBeTruthy();
-    } finally {
-      writeFactoryExportPngSpy.mockRestore();
-      exportProbe.restore();
-    }
-  });
-
-  it("switches the live dashboard to zh-CN across header, dialogs, and widgets while keeping data values stable", async () => {
-    const file = new File(["png"], "factory-import.png", { type: "image/png" });
-    const importValue = createFactoryImportValue();
-    vi.spyOn(factoryPngImportModule, "readFactoryImportPng").mockResolvedValue({
-      ok: true,
-      value: importValue,
-    });
-    const { fetchMock } = renderApp({
-      initialLocale: "en",
-      snapshot: terminalSnapshot,
-    });
-
-    fetchMock.mockImplementation(async (input: RequestInfo | URL) => {
-      const path =
-        typeof input === "string"
-          ? input
-          : input instanceof URL
-            ? `${input.pathname}${input.search}`
-            : input.url;
-
-      if (
-        path ===
-        `/factories/${DEFAULT_FACTORY_SESSION_ID}/factory/~current`
-      ) {
-        return jsonResponse(currentNamedFactoryExportResponse);
-      }
-
-      throw new Error(`unexpected fetch for ${path}`);
-    });
-
-    const englishToolbar = await screen.findByRole("region", {
-      name: "dashboard summary",
-    });
-    const languageButton = within(englishToolbar).getByRole("button", {
-      name: "Change language",
-    });
-
-    expect(
-      within(englishToolbar).getByRole("button", { name: "Export PNG" }),
-    ).toBeTruthy();
-    expect(screen.getByText("Waiting for more ticks")).toBeTruthy();
-    expect(screen.getByRole("heading", { name: "Factory graph" })).toBeTruthy();
-    expect(screen.getByRole("button", { name: "Done Story" })).toBeTruthy();
-
-    fireEvent.click(languageButton);
-    fireEvent.click(
-      screen.getByRole("menuitemradio", {
-        name: "简体中文",
-      }),
-    );
-
-    const localizedToolbar = await screen.findByRole("region", {
-      name: "仪表板概览",
-    });
-
-    expect(
-      within(localizedToolbar).getByRole("button", { name: "切换语言" }),
-    ).toBeTruthy();
-    expect(
-      within(localizedToolbar).getByRole("slider", { name: "时间线刻度" }),
-    ).toBeTruthy();
-    expect(screen.getByText("正在等待更多刻度")).toBeTruthy();
-    expect(
-      within(localizedToolbar).getByRole("button", { name: "导出 PNG" }),
-    ).toBeTruthy();
-    expect(screen.getByRole("heading", { name: "工作总计" })).toBeTruthy();
-    expect(screen.getByLabelText("已完成：1")).toBeTruthy();
-    expect(screen.getByRole("heading", { name: "工厂图" })).toBeTruthy();
-    expect(screen.getByRole("region", { name: "工作图视口" })).toBeTruthy();
-    expect(screen.getByRole("button", { name: "Done Story" })).toBeTruthy();
-
-    fireEvent.drop(
-      screen.getByRole("region", { name: "工作图视口" }),
-      createFileDropTransfer([file]),
-    );
-
-    const importDialog = await screen.findByRole("dialog", {
-      name: "检查工厂导入",
-    });
-    expect(importDialog.textContent).toContain("Dropped Factory");
-    expect(importDialog.textContent).toContain("factory-import.png");
-    expect(
-      within(importDialog).getByRole("img", {
-        name: "Dropped Factory 预览图",
-      }),
-    ).toBeTruthy();
-    expect(
-      within(importDialog).getByRole("button", { name: "启用工厂" }),
-    ).toBeTruthy();
-    expect(
-      within(importDialog).getByRole("button", { name: "取消导入" }),
-    ).toBeTruthy();
-
-    fireEvent.click(
-      within(importDialog).getByRole("button", { name: "取消导入" }),
-    );
-
-    await waitFor(() => {
-      expect(screen.queryByRole("dialog", { name: "检查工厂导入" })).toBeNull();
-    });
-
-    fireEvent.click(
-      within(localizedToolbar).getByRole("button", { name: "导出 PNG" }),
-    );
-
-    const exportDialog = await screen.findByRole("dialog", {
-      name: "导出工厂",
-    });
-    await waitFor(() => {
-      expect(
-        within(exportDialog).getByDisplayValue("semantic-workflow"),
-      ).toBeTruthy();
-    });
-    expect(within(exportDialog).getByLabelText("工厂名称")).toBeTruthy();
-    expect(within(exportDialog).getByLabelText("封面图片")).toBeTruthy();
-
-    fireEvent.click(within(exportDialog).getByRole("button", { name: "取消" }));
-
-    await waitFor(() => {
-      expect(screen.queryByRole("dialog", { name: "导出工厂" })).toBeNull();
-    });
-  });
-
-  it("falls back to English when the configured app locale is unsupported", async () => {
-    renderApp({
-      locationSearch: "?locale=fr-CA",
-      snapshot: terminalSnapshot,
-    });
-
-    expect(
-      await screen.findByRole("region", { name: "dashboard summary" }),
-    ).toBeTruthy();
-    expect(
-      screen.getByRole("button", { name: "Change language" }),
-    ).toBeTruthy();
-    expect(screen.getByText("Waiting for more ticks")).toBeTruthy();
-    expect(screen.getByRole("heading", { name: "Work totals" })).toBeTruthy();
-    expect(screen.queryByRole("button", { name: "切换语言" })).toBeNull();
-  });
-
-  it("applies the shared typography helpers to the dashboard toolbar summary shell", async () => {
-    renderApp({ snapshot: baselineSnapshot });
-
-    const heading = await screen.findByRole("heading", {
-      name: "you-agent-factory",
-    });
-    const toolbar = screen.getByRole("region", { name: "dashboard summary" });
-    const streamStatus = screen.getByRole("status", {
-      name: "you-agent-factory event stream connecting",
-    });
-    const exportButton = screen.getByRole("button", { name: "Export PNG" });
-
-    expect(heading.className).toContain(DASHBOARD_PAGE_HEADING_CLASS);
-    expect(streamStatus.className).toContain(DASHBOARD_BODY_TEXT_CLASS);
-    expect(streamStatus.className).toContain(DASHBOARD_SUPPORTING_LABELS_CLASS);
-    expect(within(toolbar).queryByText("Factory state")).toBeNull();
-    expect(
-      within(toolbar).queryByText(baselineSnapshot.factory_state),
-    ).toBeNull();
-    expect(within(toolbar).queryByText("Loading factory events...")).toBeNull();
-    expect(within(toolbar).queryByText("Export PNG")).toBeNull();
-    expect(exportButton.getAttribute("aria-haspopup")).toBe("dialog");
-  });
-
-  it("renders compact accessible stream status states without the retired toolbar labels", async () => {
-    renderApp({ snapshot: baselineSnapshot });
-
-    const toolbar = await screen.findByRole("region", {
-      name: "dashboard summary",
-    });
-
-    expect(
-      within(toolbar).getByRole("status", {
-        name: "you-agent-factory event stream connecting",
-      }),
-    ).toBeTruthy();
-    expect(within(toolbar).queryByText("Factory state")).toBeNull();
-    expect(within(toolbar).queryByText("Stream")).toBeNull();
-    expect(within(toolbar).queryByText("Loading factory events...")).toBeNull();
-    expect(within(toolbar).queryByText("Export PNG")).toBeNull();
-
-    act(() => {
-      useDashboardStreamStore.setState({
-        streamState: {
-          message: "Factory event stream connected.",
-          status: "live",
-        },
-      });
-    });
-
-    await waitFor(() => {
-      expect(
-        within(toolbar).getByRole("status", {
-          name: "you-agent-factory event stream live",
-        }),
-      ).toBeTruthy();
-    });
-    expect(
-      within(toolbar).queryByText("Factory event stream connected."),
-    ).toBeNull();
-
-    act(() => {
-      useDashboardStreamStore.setState({
-        streamState: {
-          message:
-            "Factory event stream disconnected. Showing last event state.",
-          status: "offline",
-        },
-      });
-    });
-
-    await waitFor(() => {
-      expect(
-        within(toolbar).getByRole("status", {
-          name: "you-agent-factory event stream offline",
-        }),
-      ).toBeTruthy();
-    });
-    expect(
-      within(toolbar).queryByText(
-        "Factory event stream disconnected. Showing last event state.",
-      ),
-    ).toBeNull();
-  });
-
-  it("opens the export dialog from the toolbar and dismisses it without dashboard side effects", async () => {
-    const exportProbe = installExportDownloadProbe();
-    const { fetchMock } = renderApp({
-      snapshot: baselineSnapshot,
-      timelineEvents: exportTimelineEvents,
-    });
-    fetchMock.mockResolvedValueOnce(
-      jsonResponse(currentNamedFactoryExportResponse),
-    );
-
-    try {
-      fireEvent.click(
-        await screen.findByRole("button", { name: "Export PNG" }),
-      );
-
-      const dialog = await screen.findByRole("dialog", {
-        name: "Export factory",
-      });
-      await waitFor(() => {
-        expect(
-          within(dialog).getByDisplayValue("semantic-workflow"),
-        ).toBeTruthy();
-      });
-      expect(
-        within(dialog).getByText(/without changing the live dashboard state/i),
-      ).toBeTruthy();
-      expect(
-        within(dialog).getByRole("button", { name: "Cancel" }),
-      ).toBeTruthy();
-      fireEvent.change(within(dialog).getByLabelText("Factory name"), {
-        target: { value: "Factory Poster" },
-      });
-      const imageInput = within(dialog).getByLabelText(
-        "Cover image",
-      ) as HTMLInputElement;
-      Object.defineProperty(imageInput, "files", {
-        configurable: true,
-        value: [exportImageFile()],
-      });
-      fireEvent.change(imageInput);
-      expect(
-        within(dialog).getByText("Selected image: cover.png"),
-      ).toBeTruthy();
-
-      fireEvent.click(within(dialog).getByRole("button", { name: "Cancel" }));
-
-      await waitFor(() => {
-        expect(
-          screen.queryByRole("dialog", { name: "Export factory" }),
-        ).toBeNull();
-      });
-      expect(
-        screen.getByRole("region", { name: "dashboard summary" }),
-      ).toBeTruthy();
-      expect(nonPromptTemplateFetchPaths(fetchMock)).toEqual([
-        `/factories/${DEFAULT_FACTORY_SESSION_ID}/factory/~current`,
-      ]);
-      expect(exportProbe.getDownloadedBlob()).toBeNull();
-      expect(exportProbe.getDownloadedFilename()).toBe("");
-    } finally {
-      exportProbe.restore();
-    }
-  });
-
-  it("waits for a fresh current-factory response before exporting after reopen", async () => {
-    const refreshedCurrentFactoryExportResponse = {
-      ...currentNamedFactoryExportResponse,
-      metadata: {
-        ...currentNamedFactoryExportResponse.metadata,
-        contractSource: "refetched-current-factory-api",
-      },
-      id: "authored-refetched-factory",
-      name: "imported-workflow",
-    } satisfies FactoryValue;
-    const refreshedCurrentFactoryResponse = createDeferredPromise<Response>();
-    const writeFactoryExportPngSpy = vi
-      .spyOn(factoryPngExportModule, "writeFactoryExportPng")
-      .mockResolvedValue({
-        blob: new Blob([toArrayBuffer(fromBase64(ONE_PIXEL_PNG_BASE64))], {
-          type: "image/png",
-        }),
-        metadata: {
-          ...refreshedCurrentFactoryExportResponse,
-          schemaVersion: "portos.agent-factory.png.v1",
-        },
-        ok: true,
-      });
-    const { fetchMock } = renderApp({
-      snapshot: baselineSnapshot,
-      timelineEvents: exportTimelineEvents,
-    });
-    let currentFactoryFetchCount = 0;
-
-    fetchMock.mockImplementation(async (input: RequestInfo | URL) => {
-      const path =
-        typeof input === "string"
-          ? input
-          : input instanceof URL
-            ? `${input.pathname}${input.search}`
-            : input.url;
-
-      if (
-        path !==
-        `/factories/${DEFAULT_FACTORY_SESSION_ID}/factory/~current`
-      ) {
-        throw new Error(`unexpected fetch for ${path}`);
-      }
-
-      currentFactoryFetchCount += 1;
-
-      if (currentFactoryFetchCount === 1) {
-        return jsonResponse(currentNamedFactoryExportResponse);
-      }
-
-      if (currentFactoryFetchCount === 2) {
-        return refreshedCurrentFactoryResponse.promise;
-      }
-
-      throw new Error(
-        `unexpected current factory fetch #${currentFactoryFetchCount}`,
-      );
-    });
-
-    try {
-      fireEvent.click(
-        await screen.findByRole("button", { name: "Export PNG" }),
-      );
-
-      const firstDialog = await screen.findByRole("dialog", {
-        name: "Export factory",
-      });
-      await waitFor(() => {
-        expect(
-          within(firstDialog).getByDisplayValue("semantic-workflow"),
-        ).toBeTruthy();
-      });
-      expect(
-        (
-          within(firstDialog).getByRole("button", {
-            name: "Export PNG",
-          }) as HTMLButtonElement
-        ).disabled,
-      ).toBe(false);
-
-      fireEvent.click(
-        within(firstDialog).getByRole("button", { name: "Cancel" }),
-      );
-
-      await waitFor(() => {
-        expect(
-          screen.queryByRole("dialog", { name: "Export factory" }),
-        ).toBeNull();
-      });
-
-      fireEvent.click(screen.getByRole("button", { name: "Export PNG" }));
-
-      const secondDialog = await screen.findByRole("dialog", {
-        name: "Export factory",
-      });
-      expect(
-        within(secondDialog).getByText(
-          "Loading the current authored factory definition.",
-        ),
-      ).toBeTruthy();
-      expect(
-        (
-          within(secondDialog).getByRole("button", {
-            name: "Export PNG",
-          }) as HTMLButtonElement
-        ).disabled,
-      ).toBe(true);
-      expect(writeFactoryExportPngSpy).not.toHaveBeenCalled();
-
-      await act(async () => {
-        refreshedCurrentFactoryResponse.resolve(
-          jsonResponse(refreshedCurrentFactoryExportResponse),
-        );
-        await refreshedCurrentFactoryResponse.promise;
-      });
-
-      await waitFor(() => {
-        expect(
-          within(secondDialog).getByDisplayValue("imported-workflow"),
-        ).toBeTruthy();
-      });
-      expect(
-        (
-          within(secondDialog).getByRole("button", {
-            name: "Export PNG",
-          }) as HTMLButtonElement
-        ).disabled,
-      ).toBe(false);
-
-      const imageInput = within(secondDialog).getByLabelText(
-        "Cover image",
-      ) as HTMLInputElement;
-      Object.defineProperty(imageInput, "files", {
-        configurable: true,
-        value: [exportImageFile()],
-      });
-      fireEvent.change(imageInput);
-      fireEvent.click(
-        within(secondDialog).getByRole("button", { name: "Export PNG" }),
-      );
-
-      await waitFor(() => {
-        expect(writeFactoryExportPngSpy).toHaveBeenCalledWith({
-          factory: refreshedCurrentFactoryExportResponse,
-          image: expect.any(File),
-        });
-      });
-      await waitFor(() => {
-        expect(within(secondDialog).getByRole("status").textContent).toContain(
-          "Downloaded imported-workflow.png.",
-        );
-      });
-      fireEvent.click(
-        within(secondDialog).getByRole("button", { name: "Close" }),
-      );
-      await waitFor(() => {
-        expect(
-          screen.queryByRole("dialog", { name: "Export factory" }),
-        ).toBeNull();
-      });
-    } finally {
-      writeFactoryExportPngSpy.mockRestore();
-    }
-  });
-
-  it("does not download after cancelling an export that is still in flight", async () => {
-    const exportProbe = installExportDownloadProbe();
-    const factoryPngExportModule = await import(
-      "./features/export/factory-png-export"
-    );
-    const pendingExport =
-      createDeferredPromise<
-        Awaited<ReturnType<typeof factoryPngExportModule.writeFactoryExportPng>>
-      >();
-    const writeFactoryExportPngSpy = vi
-      .spyOn(factoryPngExportModule, "writeFactoryExportPng")
-      .mockReturnValue(pendingExport.promise);
-    const { fetchMock } = renderApp({
-      snapshot: baselineSnapshot,
-      timelineEvents: exportTimelineEvents,
-    });
-    fetchMock.mockResolvedValueOnce(
-      jsonResponse(currentNamedFactoryExportResponse),
-    );
-
-    try {
-      fireEvent.click(
-        await screen.findByRole("button", { name: "Export PNG" }),
-      );
-
-      const dialog = await screen.findByRole("dialog", {
-        name: "Export factory",
-      });
-      await waitFor(() => {
-        expect(
-          (
-            within(dialog).getByRole("button", {
-              name: "Export PNG",
-            }) as HTMLButtonElement
-          ).disabled,
-        ).toBe(false);
-      });
-      fireEvent.change(within(dialog).getByLabelText("Factory name"), {
-        target: { value: "Factory Poster" },
-      });
-      const imageInput = within(dialog).getByLabelText(
-        "Cover image",
-      ) as HTMLInputElement;
-      Object.defineProperty(imageInput, "files", {
-        configurable: true,
-        value: [exportImageFile()],
-      });
-      fireEvent.change(imageInput);
-      fireEvent.click(
-        within(dialog).getByRole("button", { name: "Export PNG" }),
-      );
-
-      expect(
-        (
-          within(dialog).getByRole("button", {
-            name: "Exporting...",
-          }) as HTMLButtonElement
-        ).disabled,
-      ).toBe(true);
-
-      fireEvent.click(within(dialog).getByRole("button", { name: "Cancel" }));
-
-      await waitFor(() => {
-        expect(
-          screen.queryByRole("dialog", { name: "Export factory" }),
-        ).toBeNull();
-      });
-
-      await act(async () => {
-        pendingExport.resolve({
-          blob: new Blob([toArrayBuffer(fromBase64(ONE_PIXEL_PNG_BASE64))], {
-            type: "image/png",
-          }),
-          metadata: {
-            ...currentNamedFactoryExportResponse,
-            name: "Factory Poster",
-            schemaVersion: "portos.agent-factory.png.v1",
-          },
-          ok: true,
-        });
-        await pendingExport.promise;
-      });
-
-      expect(writeFactoryExportPngSpy).toHaveBeenCalledTimes(1);
-      expect(exportProbe.getDownloadedBlob()).toBeNull();
-      expect(exportProbe.getDownloadedFilename()).toBe("");
-    } finally {
-      writeFactoryExportPngSpy.mockRestore();
-      exportProbe.restore();
-    }
-  });
-
-  it("validates the export fields and accepts the confirmed name plus selected image", async () => {
-    const exportProbe = installExportDownloadProbe();
-    const { fetchMock } = renderApp({
-      snapshot: baselineSnapshot,
-      timelineEvents: exportTimelineEvents,
-    });
-    fetchMock.mockResolvedValueOnce(
-      jsonResponse(currentNamedFactoryExportResponse),
-    );
-
-    try {
-      fireEvent.click(
-        await screen.findByRole("button", { name: "Export PNG" }),
-      );
-
-      const dialog = await screen.findByRole("dialog", {
-        name: "Export factory",
-      });
-      await waitFor(() => {
-        expect(
-          (
-            within(dialog).getByRole("button", {
-              name: "Export PNG",
-            }) as HTMLButtonElement
-          ).disabled,
-        ).toBe(false);
-      });
-      const exportButton = within(dialog).getByRole("button", {
-        name: "Export PNG",
-      });
-      fireEvent.click(exportButton);
-
-      expect(
-        within(dialog).getByText("Choose a cover image before exporting."),
-      ).toBeTruthy();
-      expect(exportProbe.getDownloadedBlob()).toBeNull();
-
-      const nameInput = within(dialog).getByLabelText("Factory name");
-      fireEvent.change(nameInput, { target: { value: "   " } });
-      fireEvent.click(exportButton);
-
-      expect(
-        within(dialog).getByText("Enter a factory name before exporting."),
-      ).toBeTruthy();
-      expect(exportProbe.getDownloadedBlob()).toBeNull();
-
-      fireEvent.change(nameInput, { target: { value: "Factory Poster" } });
-      const imageInput = within(dialog).getByLabelText(
-        "Cover image",
-      ) as HTMLInputElement;
-      Object.defineProperty(imageInput, "files", {
-        configurable: true,
-        value: [exportImageFile()],
-      });
-      fireEvent.change(imageInput);
-
-      expect(within(dialog).getByDisplayValue("Factory Poster")).toBeTruthy();
-      expect(
-        within(dialog).getByText("Selected image: cover.png"),
-      ).toBeTruthy();
-      expect(
-        within(dialog).queryByText("Enter a factory name before exporting."),
-      ).toBeNull();
-      expect(
-        within(dialog).queryByText("Choose a cover image before exporting."),
-      ).toBeNull();
-      expect(exportProbe.getDownloadedBlob()).toBeNull();
-      expect(exportProbe.getDownloadedFilename()).toBe("");
-    } finally {
-      exportProbe.restore();
-    }
-  });
-
-  it("exports the current named-factory API payload instead of the event timeline projection", async () => {
-    const exportProbe = installExportDownloadProbe();
-    const factoryPngExportModule = await import(
-      "./features/export/factory-png-export"
-    );
-    const writeFactoryExportPngSpy = vi
-      .spyOn(factoryPngExportModule, "writeFactoryExportPng")
-      .mockResolvedValue({
-        blob: new Blob([toArrayBuffer(fromBase64(ONE_PIXEL_PNG_BASE64))], {
-          type: "image/png",
-        }),
-        metadata: {
-          ...currentNamedFactoryExportResponse,
-          name: "Factory Poster",
-          schemaVersion: "portos.agent-factory.png.v1",
-        },
-        ok: true,
-      });
-    const { fetchMock } = renderApp({
-      snapshot: baselineSnapshot,
-      timelineEvents: exportTimelineEvents,
-    });
-    fetchMock.mockResolvedValueOnce(
-      jsonResponse(currentNamedFactoryExportResponse),
-    );
-
-    try {
-      fireEvent.click(
-        await screen.findByRole("button", { name: "Export PNG" }),
-      );
-
-      const dialog = await screen.findByRole("dialog", {
-        name: "Export factory",
-      });
-      await waitFor(() => {
-        expect(
-          (
-            within(dialog).getByRole("button", {
-              name: "Export PNG",
-            }) as HTMLButtonElement
-          ).disabled,
-        ).toBe(false);
-      });
-      fireEvent.change(within(dialog).getByLabelText("Factory name"), {
-        target: { value: "Factory Poster" },
-      });
-      const imageInput = within(dialog).getByLabelText(
-        "Cover image",
-      ) as HTMLInputElement;
-      Object.defineProperty(imageInput, "files", {
-        configurable: true,
-        value: [exportImageFile()],
-      });
-      fireEvent.change(imageInput);
-      fireEvent.click(
-        within(dialog).getByRole("button", { name: "Export PNG" }),
-      );
-
-      await waitFor(() => {
-        expect(writeFactoryExportPngSpy).toHaveBeenCalledTimes(1);
-      });
-      expect(writeFactoryExportPngSpy).toHaveBeenCalledWith({
-        factory: {
-          ...currentNamedFactoryExportResponse,
-          name: "Factory Poster",
-        },
-        image: expect.any(File),
-      });
-      expect(exportProbe.getDownloadedFilename()).toBe("factory-poster.png");
-    } finally {
-      writeFactoryExportPngSpy.mockRestore();
-      exportProbe.restore();
-    }
-  });
-});
 
 describe("App timeline reconstruction flows", () => {
   registerAppDashboardTestLifecycle();
