+++ ui/src/testing/app-shell-test-utils.tsx
@@ -1,0 +1,428 @@
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { cleanup, render } from "@testing-library/react";
+import { afterEach, beforeEach, vi, type Mock } from "vitest";
+import { App } from "../App";
+import type {
+  DashboardSnapshot,
+  DashboardTrace,
+  DashboardWorkstationRequest,
+} from "../api/dashboard";
+import { DEFAULT_FACTORY_SESSION_ID } from "../api/session-routing";
+import type { FactoryEvent } from "../api/events";
+import {
+  buildDashboardSnapshotFixture,
+  mediumBranchingDashboardTopology,
+} from "../components/dashboard/fixtures";
+import { installDashboardBrowserTestShims } from "../components/dashboard/test-browser-shims";
+import { semanticWorkflowDashboardSnapshot } from "../components/dashboard/test-fixtures";
+import { reloadDashboardLayoutFromStorage } from "../features/bento";
+import { useDashboardBentoStore } from "../features/bento/state";
+import { useCurrentEditableFactoryDefinition } from "../features/current-factory-definition";
+import { resetSelectionHistoryStore } from "../features/current-selection/state";
+import {
+  createDefaultDashboardStreamState,
+  useDashboardStreamStore,
+} from "../features/dashboard/state";
+import { useDashboardSessionStore } from "../features/dashboard/state/dashboardSessionStore";
+import { useExportDialogStore } from "../features/export/state";
+import type { FactoryPngImportValue } from "../features/import";
+import type { WorldState } from "../features/timeline/state";
+import { useFactoryTimelineStore } from "../features/timeline/state";
+
+vi.mock("../features/current-factory-definition", async () => {
+  const actual = await vi.importActual("../features/current-factory-definition");
+
+  return {
+    ...actual,
+    useCurrentEditableFactoryDefinition: vi.fn(),
+  };
+});
+
+export class MockEventSource {
+  public static instances: MockEventSource[] = [];
+
+  public onerror: ((event: Event) => void) | null = null;
+  public onopen: ((event: Event) => void) | null = null;
+
+  private readonly listeners = new Map<string, EventListener[]>();
+
+  public constructor(public readonly url: string) {
+    MockEventSource.instances.push(this);
+  }
+
+  public addEventListener(type: string, listener: EventListener): void {
+    const existing = this.listeners.get(type) ?? [];
+    existing.push(listener);
+    this.listeners.set(type, existing);
+  }
+
+  public close(): void {}
+
+  public emit(type: string, data: unknown): void {
+    if (type === "snapshot") {
+      const state = useFactoryTimelineStore.getState();
+      const tracesByWorkID =
+        state.worldViewCache[state.selectedTick]?.tracesByWorkID ?? {};
+      seedTimelineSnapshot(data as DashboardSnapshot, tracesByWorkID);
+    }
+
+    const event = new MessageEvent(type, {
+      data: JSON.stringify(data),
+    });
+
+    for (const listener of this.listeners.get(type) ?? []) {
+      listener(event);
+    }
+  }
+}
+
+interface RenderAppOptions {
+  browserLanguage?: string | null;
+  browserLanguages?: readonly string[] | null;
+  initialLocale?: string | null;
+  locationSearch?: string | null;
+  snapshot: DashboardSnapshot;
+  timelineEvents?: FactoryEvent[];
+  timelineSnapshots?: DashboardSnapshot[];
+  traceFixtures?: Record<string, DashboardTrace>;
+  workstationRequestsByDispatchID?: Record<string, DashboardWorkstationRequest>;
+}
+
+type FetchMock = Mock<
+  (input: RequestInfo | URL, init?: RequestInit) => Promise<Response>
+>;
+
+interface RenderAppResult extends ReturnType<typeof render> {
+  fetchMock: FetchMock;
+}
+
+const terminalBaseSnapshot = semanticWorkflowDashboardSnapshot;
+const queryClients: QueryClient[] = [];
+let restoreBrowserTestShims: (() => void) | null = null;
+
+export const baselineSnapshot = buildDashboardSnapshotFixture(
+  mediumBranchingDashboardTopology,
+);
+
+export const terminalSnapshot = {
+  ...terminalBaseSnapshot,
+  tick_count: 4,
+  runtime: {
+    ...terminalBaseSnapshot.runtime,
+    place_occupancy_work_items_by_place_id: {
+      ...(terminalBaseSnapshot.runtime.place_occupancy_work_items_by_place_id ??
+        {}),
+      "story:blocked": [
+        {
+          display_name: "Failed Story",
+          trace_id: "trace-failed-story",
+          work_id: "work-failed-story",
+          work_type_id: "story",
+        },
+      ],
+      "story:complete": [
+        {
+          display_name: "Done Story",
+          trace_id: "trace-done-story",
+          work_id: "work-complete",
+          work_type_id: "story",
+        },
+      ],
+    },
+    place_token_counts: {
+      ...(terminalBaseSnapshot.runtime.place_token_counts ?? {}),
+      "story:blocked": 1,
+      "story:complete": 1,
+    },
+    session: {
+      ...terminalBaseSnapshot.runtime.session,
+      completed_count: 1,
+      completed_work_labels: ["Done Story"],
+      provider_sessions: [
+        ...(terminalBaseSnapshot.runtime.session.provider_sessions ?? []),
+        {
+          dispatch_id: "dispatch-complete",
+          outcome: "ACCEPTED",
+          provider_session: {
+            id: "sess-done-story",
+            kind: "session_id",
+            provider: "codex",
+          },
+          transition_id: "complete",
+          workstation_name: "Complete",
+          work_items: [
+            {
+              display_name: "Done Story",
+              trace_id: "trace-done-story",
+              work_id: "work-complete",
+              work_type_id: "story",
+            },
+          ],
+        },
+      ],
+    },
+  },
+} satisfies DashboardSnapshot;
+
+export const importedFactorySnapshot = (() => {
+  const snapshot = structuredClone(semanticWorkflowDashboardSnapshot);
+
+  snapshot.factory_state = "Imported factory active";
+  snapshot.tick_count = semanticWorkflowDashboardSnapshot.tick_count + 1;
+
+  return snapshot;
+})();
+
+function timelineSnapshot(
+  snapshot: DashboardSnapshot,
+  tracesByWorkID: Record<string, DashboardTrace> = {},
+  workstationRequestsByDispatchID: Record<
+    string,
+    DashboardWorkstationRequest
+  > = {},
+): WorldState {
+  return {
+    ...snapshot,
+    relationsByWorkID: {},
+    tracesByWorkID,
+    workstationRequestsByDispatchID,
+    workRequestsByID: {},
+  };
+}
+
+function seedTimelineSnapshot(
+  snapshot: DashboardSnapshot,
+  tracesByWorkID: Record<string, DashboardTrace> = {},
+  workstationRequestsByDispatchID: Record<
+    string,
+    DashboardWorkstationRequest
+  > = {},
+): void {
+  useFactoryTimelineStore.setState({
+    events: [],
+    latestTick: snapshot.tick_count,
+    mode: "current",
+    receivedEventIDs: [],
+    selectedTick: snapshot.tick_count,
+    worldViewCache: {
+      [snapshot.tick_count]: timelineSnapshot(
+        snapshot,
+        tracesByWorkID,
+        workstationRequestsByDispatchID,
+      ),
+    },
+  });
+}
+
+function seedTimelineSnapshots(snapshots: DashboardSnapshot[]): void {
+  const worldViewCache = Object.fromEntries(
+    snapshots.map(
+      (snapshot) =>
+        [
+          snapshot.tick_count,
+          timelineSnapshot(snapshot) satisfies WorldState,
+        ] as const,
+    ),
+  );
+  const latestTick = Math.max(
+    ...snapshots.map((snapshot) => snapshot.tick_count),
+  );
+
+  useFactoryTimelineStore.setState({
+    events: [],
+    latestTick,
+    mode: "current",
+    receivedEventIDs: [],
+    selectedTick: latestTick,
+    worldViewCache,
+  });
+}
+
+export function renderApp({
+  browserLanguage,
+  browserLanguages,
+  initialLocale,
+  locationSearch,
+  snapshot,
+  timelineEvents,
+  timelineSnapshots,
+  traceFixtures = {},
+  workstationRequestsByDispatchID = {},
+}: RenderAppOptions): RenderAppResult {
+  const queryClient = new QueryClient({
+    defaultOptions: {
+      queries: {
+        gcTime: Infinity,
+        retry: false,
+      },
+    },
+  });
+  queryClients.push(queryClient);
+
+  const fetchMock: FetchMock = vi
+    .fn()
+    .mockImplementation(async (input: RequestInfo | URL) => {
+      const path =
+        typeof input === "string"
+          ? input
+          : input instanceof URL
+            ? `${input.pathname}${input.search}`
+            : input.url;
+
+      throw new Error(`unexpected fetch for ${path}`);
+    });
+
+  vi.stubGlobal("fetch", fetchMock);
+  vi.stubGlobal("EventSource", MockEventSource);
+  reloadDashboardLayoutFromStorage();
+  if (timelineEvents) {
+    useFactoryTimelineStore.getState().replaceEvents(timelineEvents);
+  } else if (timelineSnapshots) {
+    seedTimelineSnapshots(timelineSnapshots);
+  } else {
+    seedTimelineSnapshot(
+      snapshot,
+      traceFixtures,
+      workstationRequestsByDispatchID,
+    );
+  }
+
+  const result = render(
+    <QueryClientProvider client={queryClient}>
+      <App
+        browserLanguage={browserLanguage}
+        browserLanguages={browserLanguages}
+        initialLocale={initialLocale}
+        locationSearch={locationSearch}
+      />
+    </QueryClientProvider>,
+  );
+
+  return { ...result, fetchMock };
+}
+
+function fetchCallPaths(fetchMock: ReturnType<typeof vi.fn>) {
+  return fetchMock.mock.calls.map(([input]) =>
+    typeof input === "string"
+      ? input
+      : input instanceof URL
+        ? `${input.pathname}${input.search}`
+        : input.url,
+  );
+}
+
+export function nonPromptTemplateFetchPaths(fetchMock: ReturnType<typeof vi.fn>) {
+  return fetchCallPaths(fetchMock).filter(
+    (path) =>
+      !path.includes("/prompt-template-contract") &&
+      path !== "/factory-sessions",
+  );
+}
+
+export function jsonResponse(
+  body: unknown,
+  status = 200,
+  statusText?: string,
+): Response {
+  return new Response(JSON.stringify(body), {
+    headers: {
+      "Content-Type": "application/json",
+    },
+    status,
+    statusText,
+  });
+}
+
+export function createFactoryImportValue(): FactoryPngImportValue {
+  return {
+    factory: {
+      name: "Dropped Factory",
+      workTypes: [],
+      workers: [],
+      workstations: [],
+    },
+    previewImageSrc: "blob:factory-preview",
+    revokePreviewImageSrc: vi.fn(),
+    schemaVersion: "portos.agent-factory.png.v1",
+  };
+}
+
+export function createFileDropTransfer(files: File[]): {
+  dataTransfer: {
+    dropEffect: string;
+    files: File[];
+    types: string[];
+  };
+} {
+  return {
+    dataTransfer: {
+      dropEffect: "none",
+      files,
+      types: ["Files"],
+    },
+  };
+}
+
+export function registerAppDashboardTestLifecycle(): void {
+  beforeEach(() => {
+    window.localStorage.clear();
+    MockEventSource.instances = [];
+    restoreBrowserTestShims = installDashboardBrowserTestShims();
+    resetSelectionHistoryStore();
+    useDashboardSessionStore.setState({
+      selectedSessionID: "~default",
+    });
+    vi.mocked(useCurrentEditableFactoryDefinition).mockReturnValue({
+      data: undefined,
+      error: null,
+      failureCount: 0,
+      failureReason: null,
+      fetchStatus: "idle",
+      isError: false,
+      isFetched: false,
+      isFetchedAfterMount: false,
+      isFetching: false,
+      isInitialLoading: false,
+      isLoading: false,
+      isLoadingError: false,
+      isPaused: false,
+      isPending: true,
+      isPlaceholderData: false,
+      isRefetchError: false,
+      isRefetching: false,
+      isStale: true,
+      isSuccess: false,
+      promise: Promise.resolve(undefined),
+      refetch: vi.fn(),
+      status: "pending",
+    } as never);
+  });
+
+  afterEach(() => {
+    for (const queryClient of queryClients.splice(0)) {
+      queryClient.clear();
+    }
+    cleanup();
+    useDashboardBentoStore.setState({
+      refreshToken: 0,
+      selectedTraceID: null,
+    });
+    useExportDialogStore.setState({
+      isExportDialogOpen: false,
+    });
+    useDashboardStreamStore.setState({
+      streamState: createDefaultDashboardStreamState(),
+    });
+    useDashboardSessionStore.setState({
+      selectedSessionID: "~default",
+    });
+    useFactoryTimelineStore.getState().reset();
+    resetSelectionHistoryStore();
+    restoreBrowserTestShims?.();
+    restoreBrowserTestShims = null;
+    vi.restoreAllMocks();
+    vi.unstubAllGlobals();
+  });
+}
+
+export { DEFAULT_FACTORY_SESSION_ID };
